PayloadDataLength changed from uint8_t to uint16_t. examples adapted to previous changes and updated.

+++ src/UCOPData.cpp
@@ -62,8 +62,8 @@ bool UCOPData::IsEmpty ()
 
 //--------------------------------------------------------------------
 void UCOPData::SetPayloadInfo (uint8_t* i_pPayloadBuffer,
-                               uint8_t  i_PayloadBufferLength,
-                               uint8_t  i_PayloadLength)
+                               uint16_t i_PayloadBufferLength,
+                               uint16_t i_PayloadLength)
 {
   pPayloadBuffer      = i_pPayloadBuffer;
   PayloadBufferLength = i_PayloadBufferLength;
